Fixed 32-bit code assumption. Fixes #3682.
When allocating an Array of size 2**29, Rubinius would allocate a Tuple big
enough to hold 2**29 == 536870912 elements. On a 64bit arch, each of these
elements require sizeof(intptr_t) == 8 bytes. A Tuple memory header is 4 *
sizeof(intptr_t) == 32 bytes. So, we have 2**29*8+32 == 4294967328. It's a big
number, but not so big that the 64bit arch can't easily handle it.

Unfortunately, down in the managed memory code, there was an explicit uint32_t
data type. What happens when 4294967328 == 0x100000020 bytes get put into
32bits? 0x100000020 & 0xffffffff == 32. Those 2**29 elements were being
crammed into 32 bytes. Computers may be magic, but even they can only take so
much.

+++ machine/memory/immix_collector.cpp
@@ -130,7 +130,7 @@ namespace memory {
     return obj->in_immix_p();
   }
 
-  Object* ImmixGC::allocate(uint32_t bytes, bool& collect_now) {
+  Object* ImmixGC::allocate(size_t bytes, bool& collect_now) {
     if(bytes > cMaxObjectSize) return 0;
 
     Object* obj = allocator_.allocate(bytes, collect_now).as<Object>();
@@ -142,7 +142,7 @@ namespace memory {
     return obj;
   }
 
-  Object* ImmixGC::move_object(Object* orig, uint32_t bytes, bool& collect_now) {
+  Object* ImmixGC::move_object(Object* orig, size_t bytes, bool& collect_now) {
     if(bytes > cMaxObjectSize) return 0;
 
     Object* obj = allocator_.allocate(bytes, collect_now).as<Object>();
